Refactor UI-related classes to a new Module.UI namespace
 all classes and components of Modules that are directly related to the user interface (UI) have been moved to a new namespace: Module.UI.
