+++ FoliCon/App.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using FoliCon.Modules.Configuration;
+using FoliCon.Modules.UI;
 using FoliCon.Modules.utils;
 using FoliCon.ViewModels;
 using NLog;
