+++ FoliCon/Modules/utils/FileUtils.cs
@@ -1,6 +1,7 @@
 ﻿using FoliCon.Modules.Configuration;
 using FoliCon.Modules.IGDB;
 using FoliCon.Modules.TMDB;
+using FoliCon.Modules.UI;
 using NLog;
 using Collection = TMDbLib.Objects.Collections.Collection;
 using Logger = NLog.Logger;
