+++ FoliCon/ViewModels/ProSearchResultViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using FoliCon.Modules.Extension;
+using FoliCon.Modules.UI;
 using FoliCon.Modules.utils;
 using NLog;
 using Logger = NLog.Logger;
