+++ FoliCon/Views/CustomIconControl.xaml
@@ -12,6 +12,7 @@
              xmlns:langs="clr-namespace:FoliCon.Properties.Langs"
              xmlns:extension="clr-namespace:FoliCon.Modules.Extension"
              xmlns:utils="clr-namespace:FoliCon.Modules.utils"
+             xmlns:ui="clr-namespace:FoliCon.Modules.UI"
              mc:Ignorable="d"
              d:DataContext="{d:DesignInstance viewModels:CustomIconControlViewModel}"
              Background="{DynamicResource RegionBrush}"
@@ -53,20 +54,20 @@
             </Grid.ColumnDefinitions>
             <Button x:Name="FolderButton" Grid.Row="0" Grid.Column="0" Margin="10,10,10,10" Content="{extension:Lang Key={x:Static langs:LangKeys.BrowseFolderDropHere}}"
                     HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Height="auto"
-                    Command="{Binding LoadDirectory }"  modules:FolderDragDropHelper.IsFileDragDropEnabled="True"
-            modules:FolderDragDropHelper.FileDragDropTarget="{Binding}"/>
+                    Command="{Binding LoadDirectory }"  ui:FolderDragDropHelper.IsFileDragDropEnabled="True"
+            ui:FolderDragDropHelper.FileDragDropTarget="{Binding}"/>
             <Button x:Name="IconButton" Grid.Row="0" Grid.Column="1" Margin="10,10,10,10" Content="{extension:Lang Key={x:Static langs:LangKeys.BrowseIconsDropHere}}"
                     HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Height="auto"
-                    Command="{Binding LoadIcons}"  modules:FolderDragDropHelper.IsFileDragDropEnabled="True"
-            modules:FolderDragDropHelper.FileDragDropTarget="{Binding}" />
+                    Command="{Binding LoadIcons}"  ui:FolderDragDropHelper.IsFileDragDropEnabled="True"
+            ui:FolderDragDropHelper.FileDragDropTarget="{Binding}" />
             <TextBlock Grid.Row="1" FontSize="20" Grid.Column="0" Text="{extension:Lang Key={x:Static langs:LangKeys.Folders}}" HorizontalAlignment="Center" Foreground="{DynamicResource PrimaryTextBrush}"/>
             <TextBlock Grid.Row="1" FontSize="20" Grid.Column="1" Text="{extension:Lang Key={x:Static langs:LangKeys.Icons}}" HorizontalAlignment="Center" Foreground="{DynamicResource PrimaryTextBrush}"/>
             <ListBox x:Name="FoldersList" Grid.Row="2" Margin="10,0,10,0" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" AllowDrop="True"
                      Grid.Column="0" IsManipulationEnabled="True" dd:DragDrop.IsDragSource="True" dd:DragDrop.DropHandler="{Binding ReOrderDropHandler}"
                      dd:DragDrop.IsDropTarget="True" SelectionMode="Extended"
                      ItemsSource="{Binding Path=Directories, UpdateSourceTrigger=PropertyChanged}"
-                     modules:FolderDragDropHelper.IsFileDragDropEnabled="True"
-                     modules:FolderDragDropHelper.FileDragDropTarget="{Binding}" hc:Empty.ShowEmpty="True">
+                     ui:FolderDragDropHelper.IsFileDragDropEnabled="True"
+                     ui:FolderDragDropHelper.FileDragDropTarget="{Binding}" hc:Empty.ShowEmpty="True">
                 <i:Interaction.Triggers>
                     <i:KeyTrigger Key="Delete">
                         <i:InvokeCommandAction CommandParameter="{Binding ElementName=FoldersList, Path=SelectedItems}" Command="{Binding KeyPressFolderList}" />
@@ -77,8 +78,8 @@
                      dd:DragDrop.IsDragSource="True" dd:DragDrop.IsDropTarget="True" SelectionMode="Extended"  
                      dd:DragDrop.DropHandler="{Binding ReOrderDropHandler}"
                      ItemsSource="{Binding Path=Icons, UpdateSourceTrigger=PropertyChanged}"
-                     modules:FolderDragDropHelper.IsFileDragDropEnabled="True"
-                     modules:FolderDragDropHelper.FileDragDropTarget="{Binding}" hc:Empty.ShowEmpty="True">
+                     ui:FolderDragDropHelper.IsFileDragDropEnabled="True"
+                     ui:FolderDragDropHelper.FileDragDropTarget="{Binding}" hc:Empty.ShowEmpty="True">
                 <i:Interaction.Triggers>
                     <i:KeyTrigger Key="Delete">
                         <i:InvokeCommandAction CommandParameter="{Binding ElementName=IconsList, Path=SelectedItems}" Command="{Binding KeyPressIconsList}" />
