+++ FoliCon/Views/MainWindow.xaml
@@ -12,13 +12,14 @@
         xmlns:convertor="clr-namespace:FoliCon.Modules.Convertor"
         xmlns:extension="clr-namespace:FoliCon.Modules.Extension"
         xmlns:utils="clr-namespace:FoliCon.Modules.utils"
+        xmlns:ui="clr-namespace:FoliCon.Modules.UI"
         mc:Ignorable="d"
         prism:ViewModelLocator.AutoWireViewModel="True"
         Title="{Binding Title}" Height="714" Width="851" MinWidth="850" MinHeight="630"
         WindowStartupLocation="CenterScreen" hc:WindowAttach.IsDragElement="True"
            Icon="/Resources/folicon Icon.ico"
-        modules:FolderDragDropHelper.IsFileDragDropEnabled="True"
-        modules:FolderDragDropHelper.FileDragDropTarget="{Binding}" AllowDrop="True"
+        ui:FolderDragDropHelper.IsFileDragDropEnabled="True"
+        ui:FolderDragDropHelper.FileDragDropTarget="{Binding}" AllowDrop="True"
         d:DataContext="{d:DesignInstance viewModels:MainWindowViewModel}">
 
     <Window.Resources>
