+++ FoliCon/Views/PosterPicker.xaml
@@ -10,6 +10,7 @@
              xmlns:b="http://schemas.microsoft.com/xaml/behaviors"
              xmlns:modules="clr-namespace:FoliCon.Modules"
              xmlns:extension="clr-namespace:FoliCon.Modules.Extension"
+             xmlns:ui="clr-namespace:FoliCon.Modules.UI"
              mc:Ignorable="d"
              prism:ViewModelLocator.AutoWireViewModel="True"
              d:DataContext="{d:DesignInstance viewModels:PosterPickerViewModel }">
@@ -49,7 +50,7 @@
                             <Image Source="{Binding Image}" Tag="{Binding Url}" Height="128" Width="128"
                                    RenderOptions.BitmapScalingMode="HighQuality" Margin="5, 5, 5, 5">
                                 <b:Interaction.Behaviors>
-                                    <modules:ClickBehavior CommandParameter="{Binding Path=Tag, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type Image}}}" DoubleClickCommand="{Binding RelativeSource={RelativeSource FindAncestor,
+                                    <ui:ClickBehavior CommandParameter="{Binding Path=Tag, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type Image}}}" DoubleClickCommand="{Binding RelativeSource={RelativeSource FindAncestor,
                                                         AncestorType={x:Type Window}}, Path=DataContext.(viewModels:PosterPickerViewModel.PickCommand)}" 
                                                          ClickCommand="{Binding RelativeSource={RelativeSource FindAncestor,
                                                         AncestorType={x:Type Window}}, Path=DataContext.(viewModels:PosterPickerViewModel.OpenImageCommand)}"/>
